server/requirements.txt
open-interpreter
gtts
playsound
pyobjc
pyaudio